Update Statistics.java
cast int to double

+++ gunslinger/sim/Statistics.java
@@ -267,7 +267,7 @@ public class Statistics
                 for (int team = 0; team < nplayers; team++) {
                 	double avgAliveFriends = 0;
                 	double avgDeadEnemies = 0;
-                	double avgScore = totalScores[team] / games;
+                	double avgScore = (double)totalScores[team] / (double)games;
                 	double avgAlive = 0;
                 	for (int score = 0; score < nfriends+nenemies+1+1; score++) {
                 		TableRow tr = new TableRow();
